Finish creating Receipt and receipt detail, minor refactoring

+++ src/model/Order.java
@@ -12,7 +12,7 @@ public class Order {
 	private Vector<OrderItem> orderItems;
 	private String orderStatus;
 	private Date orderDate;
-	private Integer orderTotal;
+	private Double orderTotal;
 	
 	public Order(Integer orderId, User orderUser, Vector<OrderItem> orderItems, String orderStatus, Date orderDate) {
 		super();
@@ -21,7 +21,7 @@ public class Order {
 		this.orderItems = orderItems;
 		this.orderStatus = orderStatus;
 		this.orderDate = orderDate;
-		orderTotal = 0;
+		orderTotal = (double) 0;
 	}
 	
 	public Integer getOrderId() {
@@ -65,7 +65,7 @@ public class Order {
 		this.orderDate = orderDate;
 	}
 
-	public Integer getOrderTotal() {
+	public Double getOrderTotal() {
 		return orderTotal;
 	}
